+++ tests/tools_tests/test.sh
@@ -16,6 +16,16 @@ for file in ppx/*.res; do
   fi
 done
 
+# Test format-docstrings command
+for file in src/docstrings-format/*.{res,resi,md}; do
+  output="src/expected/$(basename $file).expected"
+  ../../_build/install/default/bin/rescript-tools format-codeblocks "$file" --stdout > $output
+  # # CI. We use LF, and the CI OCaml fork prints CRLF. Convert.
+  if [ "$RUNNER_OS" == "Windows" ]; then
+    perl -pi -e 's/\r\n/\n/g' -- $output
+  fi
+done
+
 warningYellow='\033[0;33m'
 successGreen='\033[0;32m'
 reset='\033[0m'