+++ tools/src/tools.ml
@@ -676,3 +676,247 @@ let extractEmbedded ~extensionPoints ~filename =
              ("loc", Some (Analysis.Utils.cmtLocToRange loc |> stringifyRange));
            ])
   |> List.rev |> array
+
+module FormatCodeblocks = struct
+  module Transform = struct
+    type transform = AssertEqualFnToEquals  (** assertEqual(a, b) -> a == b *)
+
+    (** Transforms for the code blocks themselves. *)
+    let transform ~transforms ast =
+      match transforms with
+      | [] -> ast
+      | transforms ->
+        let hasTransform transform = transforms |> List.mem transform in
+        let mapper =
+          {
+            Ast_mapper.default_mapper with
+            expr =
+              (fun mapper exp ->
+                match exp.pexp_desc with
+                | Pexp_apply
+                    {
+                      funct =
+                        {
+                          pexp_desc =
+                            Pexp_ident
+                              ({txt = Lident "assertEqual"} as identTxt);
+                        } as ident;
+                      partial = false;
+                      args = [(Nolabel, _arg1); (Nolabel, _arg2)] as args;
+                    }
+                  when hasTransform AssertEqualFnToEquals ->
+                  {
+                    exp with
+                    pexp_desc =
+                      Pexp_apply
+                        {
+                          funct =
+                            {
+                              ident with
+                              pexp_desc =
+                                Pexp_ident {identTxt with txt = Lident "=="};
+                            };
+                          args;
+                          partial = false;
+                          transformed_jsx = false;
+                        };
+                  }
+                | _ -> Ast_mapper.default_mapper.expr mapper exp);
+          }
+        in
+        mapper.structure mapper ast
+  end
+
+  let isResLang lang =
+    match String.lowercase_ascii lang with
+    | "res" | "rescript" | "resi" -> true
+    | lang ->
+      (* Cover ```res example, and similar *)
+      String.starts_with ~prefix:"res " lang
+      || String.starts_with ~prefix:"rescript " lang
+      || String.starts_with ~prefix:"resi " lang
+
+  let formatRescriptCodeBlocks content ~transformAssertEqual ~displayFilename
+      ~addError ~markdownBlockStartLine =
+    (* Detect ReScript code blocks. *)
+    let hadCodeBlocks = ref false in
+    let block _m = function
+      | Cmarkit.Block.Code_block (codeBlock, meta) -> (
+        match Cmarkit.Block.Code_block.info_string codeBlock with
+        | Some ((lang, _) as info_string) when isResLang lang ->
+          hadCodeBlocks := true;
+
+          let currentLine =
+            meta |> Cmarkit.Meta.textloc |> Cmarkit.Textloc.first_line |> fst
+          in
+          (* Account for 0-based line numbers *)
+          let currentLine = currentLine + 1 in
+          let layout = Cmarkit.Block.Code_block.layout codeBlock in
+          let code = Cmarkit.Block.Code_block.code codeBlock in
+          let codeText =
+            code |> List.map Cmarkit.Block_line.to_string |> String.concat "\n"
+          in
+
+          let n = List.length code in
+          let newlinesNeeded =
+            max 0 (markdownBlockStartLine + currentLine - n)
+          in
+          let codeWithOffset = String.make newlinesNeeded '\n' ^ codeText in
+          let reportParseError diagnostics =
+            let buf = Buffer.create 1000 in
+            let formatter = Format.formatter_of_buffer buf in
+            Res_diagnostics.print_report ~formatter
+              ~custom_intro:(Some "Syntax error in code block in docstring")
+              diagnostics codeWithOffset;
+            addError (Buffer.contents buf)
+          in
+          let formattedCode =
+            if lang |> String.split_on_char ' ' |> List.hd = "resi" then
+              let {Res_driver.parsetree; comments; invalid; diagnostics} =
+                Res_driver.parse_interface_from_source ~for_printer:true
+                  ~display_filename:displayFilename ~source:codeWithOffset
+              in
+              if invalid then (
+                reportParseError diagnostics;
+                code)
+              else
+                Res_printer.print_interface parsetree ~comments
+                |> String.trim |> Cmarkit.Block_line.list_of_string
+            else
+              let {Res_driver.parsetree; comments; invalid; diagnostics} =
+                Res_driver.parse_implementation_from_source ~for_printer:true
+                  ~display_filename:displayFilename ~source:codeWithOffset
+              in
+              if invalid then (
+                reportParseError diagnostics;
+                code)
+              else
+                let parsetree =
+                  if transformAssertEqual then
+                    Transform.transform ~transforms:[AssertEqualFnToEquals]
+                      parsetree
+                  else parsetree
+                in
+                Res_printer.print_implementation parsetree ~comments
+                |> String.trim |> Cmarkit.Block_line.list_of_string
+          in
+
+          let mappedCodeBlock =
+            Cmarkit.Block.Code_block.make ~layout ~info_string formattedCode
+          in
+          Cmarkit.Mapper.ret (Cmarkit.Block.Code_block (mappedCodeBlock, meta))
+        | _ -> Cmarkit.Mapper.default)
+      | _ -> Cmarkit.Mapper.default
+    in
+    let mapper = Cmarkit.Mapper.make ~block () in
+    let newContent =
+      content
+      |> Cmarkit.Doc.of_string ~locs:true
+      |> Cmarkit.Mapper.map_doc mapper
+      |> Cmarkit_commonmark.of_doc
+    in
+    (newContent, !hadCodeBlocks)
+
+  let formatCodeBlocksInFile ~outputMode ~transformAssertEqual ~entryPointFile =
+    let path =
+      match Filename.is_relative entryPointFile with
+      | true -> Unix.realpath entryPointFile
+      | false -> entryPointFile
+    in
+    let errors = ref [] in
+    let addError error = errors := error :: !errors in
+
+    let makeMapper ~transformAssertEqual ~displayFilename =
+      {
+        Ast_mapper.default_mapper with
+        attribute =
+          (fun mapper ((name, payload) as attr) ->
+            match (name, Ast_payload.is_single_string payload, payload) with
+            | ( {txt = "res.doc"},
+                Some (contents, None),
+                PStr [{pstr_desc = Pstr_eval ({pexp_loc}, _)}] ) ->
+              let formattedContents, hadCodeBlocks =
+                formatRescriptCodeBlocks ~transformAssertEqual ~addError
+                  ~displayFilename
+                  ~markdownBlockStartLine:pexp_loc.loc_start.pos_lnum contents
+              in
+              if hadCodeBlocks && formattedContents <> contents then
+                ( name,
+                  PStr
+                    [
+                      Ast_helper.Str.eval
+                        (Ast_helper.Exp.constant
+                           (Pconst_string (formattedContents, None)));
+                    ] )
+              else attr
+            | _ -> Ast_mapper.default_mapper.attribute mapper attr);
+      }
+    in
+    let content =
+      if Filename.check_suffix path ".md" then
+        let content =
+          let ic = open_in path in
+          let n = in_channel_length ic in
+          let s = Bytes.create n in
+          really_input ic s 0 n;
+          close_in ic;
+          Bytes.to_string s
+        in
+        let displayFilename = Filename.basename path in
+        let formattedContents, hadCodeBlocks =
+          formatRescriptCodeBlocks ~transformAssertEqual ~addError
+            ~displayFilename ~markdownBlockStartLine:1 content
+        in
+        if hadCodeBlocks && formattedContents <> content then
+          Ok (formattedContents, content)
+        else Ok (content, content)
+      else if Filename.check_suffix path ".res" then
+        let parser =
+          Res_driver.parsing_engine.parse_implementation ~for_printer:true
+        in
+        let {Res_driver.parsetree = structure; comments; source; filename} =
+          parser ~filename:path
+        in
+        let filename = Filename.basename filename in
+        let mapper =
+          makeMapper ~transformAssertEqual ~displayFilename:filename
+        in
+        let astMapped = mapper.structure mapper structure in
+        Ok (Res_printer.print_implementation astMapped ~comments, source)
+      else if Filename.check_suffix path ".resi" then
+        let parser =
+          Res_driver.parsing_engine.parse_interface ~for_printer:true
+        in
+        let {Res_driver.parsetree = signature; comments; source; filename} =
+          parser ~filename:path
+        in
+        let mapper =
+          makeMapper ~transformAssertEqual ~displayFilename:filename
+        in
+        let astMapped = mapper.signature mapper signature in
+        Ok (Res_printer.print_interface astMapped ~comments, source)
+      else
+        Error
+          (Printf.sprintf
+             "File extension not supported. This command accepts .res, .resi, \
+              and .md files")
+    in
+    match content with
+    | Error e -> Error e
+    | Ok (formatted_content, source) ->
+      let errors = !errors in
+      if List.length errors > 0 then (
+        errors |> List.rev |> String.concat "\n" |> print_endline;
+        Error
+          (Printf.sprintf "%s: Error formatting docstrings."
+             (Filename.basename path)))
+      else if formatted_content <> source then (
+        match outputMode with
+        | `Stdout -> Ok formatted_content
+        | `File ->
+          let oc = open_out path in
+          Printf.fprintf oc "%s" formatted_content;
+          close_out oc;
+          Ok (Filename.basename path ^ ": formatted successfully"))
+      else Ok (Filename.basename path ^ ": needed no formatting")
+end